requirements.txt
streamlit
websocket-client
pandas
numpy
plotly